+++ sdk/appcenter-distribute/src/main/java/com/microsoft/appcenter/distribute/InstallerUtils.java
@@ -92,6 +92,15 @@ public class InstallerUtils {
     }
 
     /**
+     * Add new stores to local stores list.
+     *
+     * @param stores list of stores.
+     */
+    public static void addLocalStores(Set<String> stores) {
+        LOCAL_STORES.addAll(stores);
+    }
+
+    /**
      * Check whether user enabled installation via unknown sources.
      *
      * @param context any context.