+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/utils/crypto/CryptoUtils.java
@@ -11,6 +11,7 @@ import android.content.Context;
 import android.os.Build;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import androidx.annotation.VisibleForTesting;
 import android.util.Base64;
 
@@ -68,7 +69,12 @@ public class CryptoUtils {
 
         @Override
         public ICipher getCipher(String transformation, String provider) throws Exception {
-            final Cipher cipher = Cipher.getInstance(transformation, provider);
+            final Cipher cipher;
+            if (provider != null) {
+                cipher = Cipher.getInstance(transformation, provider);
+            } else {
+                cipher = Cipher.getInstance(transformation);
+            }
             return new ICipher() {
 
                 @Override
@@ -176,6 +182,7 @@ public class CryptoUtils {
         /* We have to use AES to be compliant but it's available only after Android M. */
         if (keyStore != null && apiLevel >= Build.VERSION_CODES.M) {
             try {
+                registerHandler(new CryptoAesAndEtmHandler());
                 registerHandler(new CryptoAesHandler());
             } catch (Exception e) {
                 AppCenterLog.error(LOG_TAG, "Cannot use modern encryption on this device.");